requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-groq==0.1.6 
transformers==4.43.2 
sentence-transformers==3.0.1 
python-dotenv
pypdf
langchain_google_genai